+++ csharp-netcore-modified/src/BungieNetApi/Model/DestinyComponentsRecordsDestinyCharacterRecordsComponent.cs
@@ -38,7 +38,7 @@ namespace BungieNetApi.Model
         /// <param name="records">records.</param>
         /// <param name="recordCategoriesRootNodeHash">The hash for the root presentation node definition of Triumph categories..</param>
         /// <param name="recordSealsRootNodeHash">The hash for the root presentation node definition of Triumph Seals..</param>
-        public DestinyComponentsRecordsDestinyCharacterRecordsComponent(List<int> featuredRecordHashes = default(List<int>), Dictionary<string, DestinyComponentsRecordsDestinyRecordComponent> records = default(Dictionary<string, DestinyComponentsRecordsDestinyRecordComponent>), int recordCategoriesRootNodeHash = default(int), int recordSealsRootNodeHash = default(int))
+        public DestinyComponentsRecordsDestinyCharacterRecordsComponent(List<long> featuredRecordHashes = default(List<long>), Dictionary<string, DestinyComponentsRecordsDestinyRecordComponent> records = default(Dictionary<string, DestinyComponentsRecordsDestinyRecordComponent>), long recordCategoriesRootNodeHash = default(long), long recordSealsRootNodeHash = default(long))
         {
             this.FeaturedRecordHashes = featuredRecordHashes;
             this.Records = records;
@@ -50,7 +50,7 @@ namespace BungieNetApi.Model
         /// Gets or Sets FeaturedRecordHashes
         /// </summary>
         [DataMember(Name="featuredRecordHashes", EmitDefaultValue=false)]
-        public List<int> FeaturedRecordHashes { get; set; }
+        public List<long> FeaturedRecordHashes { get; set; }
 
         /// <summary>
         /// Gets or Sets Records
@@ -63,14 +63,14 @@ namespace BungieNetApi.Model
         /// </summary>
         /// <value>The hash for the root presentation node definition of Triumph categories.</value>
         [DataMember(Name="recordCategoriesRootNodeHash", EmitDefaultValue=false)]
-        public int RecordCategoriesRootNodeHash { get; set; }
+        public long RecordCategoriesRootNodeHash { get; set; }
 
         /// <summary>
         /// The hash for the root presentation node definition of Triumph Seals.
         /// </summary>
         /// <value>The hash for the root presentation node definition of Triumph Seals.</value>
         [DataMember(Name="recordSealsRootNodeHash", EmitDefaultValue=false)]
-        public int RecordSealsRootNodeHash { get; set; }
+        public long RecordSealsRootNodeHash { get; set; }
 
         /// <summary>
         /// Returns the string presentation of the object
